+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Fixed
+ - Resolve broken `File.ReadLine` native
+ - Handle natives' "bytes written" counts when null terminators are involved
 
 
 ## [0.2.0] — 2023-02-22